+++ app/Http/Requests/Frontend/User/UpdatePasswordRequest.php
@@ -2,10 +2,9 @@
 
 namespace App\Http\Requests\Frontend\User;
 
-use App\Rules\Auth\ChangePassword;
 use App\Rules\Auth\UnusedPassword;
 use Illuminate\Foundation\Http\FormRequest;
-use DivineOmega\LaravelPasswordExposedValidationRule\PasswordExposed;
+use LangleyFoxall\LaravelNISTPasswordRules\PasswordRules;
 
 /**
  * Class UpdatePasswordRequest.
@@ -31,13 +30,15 @@ class UpdatePasswordRequest extends FormRequest
     {
         return [
             'old_password' => ['required'],
-            'password' => [
-                'required',
-                'confirmed',
-                new ChangePassword(),
-                new PasswordExposed(),
-                new UnusedPassword($this->user()),
-            ],
+            'password' => array_merge(
+                [
+                    new UnusedPassword($this->user()),
+                ],
+                PasswordRules::changePassword(
+                    $this->email,
+                    config('access.users.password_history') ? 'old_password' : null
+                )
+            ),
         ];
     }
 }