+++ config/services.php
@@ -73,17 +73,7 @@ return [
         'client_id' => env('GOOGLE_CLIENT_ID'),
         'client_secret' => env('GOOGLE_CLIENT_SECRET'),
         'redirect' => env('GOOGLE_REDIRECT'),
-
-        /*
-         * Only allows google to grab email address
-         * Default scopes array also has: 'https://www.googleapis.com/auth/plus.login'
-         * https://medium.com/@njovin/fixing-laravel-socialite-s-google-permissions-2b0ef8c18205
-         */
-        'scopes' => [
-            'https://www.googleapis.com/auth/plus.me',
-            'https://www.googleapis.com/auth/plus.profile.emails.read',
-        ],
-
+        'scopes' => [],
         'with' => [],
     ],