+++ resources/lang/az/buttons.php
@@ -1,0 +1,55 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Buttons Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used in buttons throughout the system.
+    | Regardless where it is placed, a button can be listed here so it is easily
+    | found in a intuitive way.
+    |
+    */
+
+    'backend' => [
+        'access' => [
+            'users' => [
+                'activate' => 'Aktiv et',
+                'change_password' => 'Şifrəni dəyiş',
+                'clear_session' => 'Sessiyanı təmizlə',
+                'confirm' => 'Təsdiqlə',
+                'deactivate' => 'Deaktiv et',
+                'delete_permanently' => 'Birdəfəlik sil',
+                'login_as' => ':user kimi daxil ol',
+                'resend_email' => 'Təsdiqləmə emailini yenidən göndər',
+                'restore_user' => 'İstifadəçini yenidən yüklə',
+                'unconfirm' => 'Təsdiqləmə',
+                'unlink' => 'Əlaqəni sil',
+            ],
+        ],
+    ],
+
+    'emails' => [
+        'auth' => [
+            'confirm_account' => 'Hesab Təsdiqi',
+            'reset_password' => 'Şifrəni yenilə',
+        ],
+    ],
+
+    'general' => [
+        'cancel' => 'İmtina et',
+        'continue' => 'Davam et',
+
+        'crud' => [
+            'create' => 'Yarat',
+            'delete' => 'Sil',
+            'edit' => 'Redaktə',
+            'update' => 'Yenilə',
+            'view' => 'Bax',
+        ],
+
+        'save' => 'Yadda Saxla',
+        'view' => 'Bax',
+    ],
+];