+++ resources/lang/az/navs.php
@@ -1,0 +1,35 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Navs Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used in menu items throughout the system.
+    | Regardless where it is placed, a menu item can be listed here so it is easily
+    | found in a intuitive way.
+    |
+    */
+
+    'general' => [
+        'home' => 'Ana Səhifə',
+        'logout' => 'Çıxış',
+    ],
+
+    'frontend' => [
+        'contact' => 'Əlaqə',
+        'dashboard' => 'Dashboard',
+        'login' => 'Daxil ol',
+        'macros' => 'Makrolar',
+        'register' => 'Qeydiyyat',
+
+        'user' => [
+            'account' => 'Mənim Hesabım',
+            'administration' => 'İdarə heyəti',
+            'change_password' => 'Şifrəni Dəyiş',
+            'my_information' => 'Mənim Məlumatlarım',
+            'profile' => 'Profil',
+        ],
+    ],
+];