+++ resources/lang/az/passwords.php
@@ -1,0 +1,20 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Password Reset Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are the default lines which match reasons
+    | that are given by the password broker for a password update attempt
+    | has failed, such as for an invalid token or invalid new password.
+    |
+    */
+
+    'password' => 'Şifrə ən azı 6 simvol olmalıdır və təkrar şifrən ilə eyni olmalıdır.',
+    'reset' => 'Şifrəniz sıfırlandı!',
+    'sent' => 'Elektron poçt ünvanınıza şifrənin sıfırlanması üçün məktub göndərildi',
+    'token' => 'Şifrəni sıfırlamaq üçün istifadə etdiyiniz token keçərsizdir.',
+    'user' => ' Qeyd etdiyiniz Elektron poçt ünvanına aid istifadəçi tapa bilmədik',
+];