+++ resources/lang/nl/auth.php
@@ -14,7 +14,6 @@ return [
 
     'failed' => 'Toegangsgegevens niet gevonden',
     'general_error' => 'Je hebt niet de rechten om dat te doen.',
-    'password_rules' => 'Je wachtwoord moet minstens 8 tekens lang zijn, moet minstens 1 hoofdletter, 1 kleine letter en 1 nummer bevatten.',
     'password_used' => 'Je kan geen wachtwoord instellen dat je al eerder gebruikte.',
     'socialite' => [
         'unacceptable' => ':provider is niet een geaccepteerd login type.',