+++ CHANGELOG.md
@@ -3,6 +3,18 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [5.3.8] - 2019-08-21
+### Added
+- Added Azerbaijan language (https://github.com/rappasoft/laravel-boilerplate/pull/1254)
+- Added NIST Password Rules (https://github.com/rappasoft/laravel-boilerplate/pull/1258)
+
+### Changed
+- Assign all permissions to the Admin role without the need to explicitly assign the roles/permissions to the user. (https://github.com/rappasoft/laravel-boilerplate/pull/1227)
+
+### Removed
+- Removed default Google scopes (https://github.com/rappasoft/laravel-boilerplate/pull/1253/files)
+- Removed ChangePassword rule as the new NIST rules cover it
+
 ## 5.3.7 - 2019-08-21
 ### Added
 - Actual changelog
@@ -95,4 +107,5 @@ All notable changes to this project will be documented in this file.
 - Fix yarn tests
 - Fix: Socially logged in users get assigned the default role
 
-[Unreleased]: https://github.com/rappasoft/laravel-boilerplate/compare/v5.3.7...HEAD
+[Unreleased]: https://github.com/rappasoft/laravel-boilerplate/compare/v5.3.8...HEAD
+[5.3.8]: https://github.com/olivierlacan/keep-a-changelog/compare/v5.3.7...v5.3.8